requirements.txt
pandas~=2.2
requests~=2.32
beautifulsoup4~=4.12
psycopg2-binary~=2.9
sqlalchemy~=2.0
pytest~=8.0
pytest-cov~=6.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-api-python-client~=2.152
greenlet==3.2.4
gspread==6.2.1
oauth2client==4.1.3
oauthlib==3.3.1